refactor: 폴더구조 반영
