+++ src/main/java/com/dailyemotion/common/config/FirebaseConfig.java
@@ -1,4 +1,4 @@
-package com.dailyemotion.dailyemotion.config;
+package com.dailyemotion.common.config;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
